+++ src/styling.hh
@@ -220,6 +220,7 @@ struct lnav_theme {
     positioned_property<style_config> lt_style_snippet_border;
     positioned_property<style_config> lt_style_indent_guide;
     positioned_property<style_config> lt_style_timeline_bar;
+    positioned_property<style_config> lt_style_context_line;
     std::map<log_level_t, positioned_property<style_config>> lt_level_styles;
     std::map<std::string, highlighter_config> lt_highlights;
 };