+++ test/test_cmds.sh
@@ -722,3 +722,7 @@ run_cap_test ${lnav_test} -n \
 run_cap_test ${lnav_test} -n \
     -c ':echo Hello, $XYZ!' \
     ${test_dir}/logfile_access_log.0
+
+run_cap_test ${lnav_test} -n \
+    -H \
+    -c ':filter-context 4'